refactor: split bus_bot.py into config, helpers, and handler modules
Break the 540-line monolith into focused modules for easier navigation
and maintenance. No logic changes — only file moves and import updates.
